+++ CHANGELOG.md
@@ -27,7 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-nil.
+- Remove `Session` usage from `CandidateResult` and the unnecessary `CandidateTuple` object [#152](https://github.com/Shopify/atlas_engine/pull/152)
 
 ---